Refactor : interception route를 위한 폴더 구조 변경
